+++ iOS/G3MiOSSDK/Commons/Mesh/AbstractMesh.cpp
@@ -216,11 +216,11 @@ Mesh* AbstractMesh::createNormalsMesh() const{
                                             false,
                                             _center,
                                             _vertices,
-                                            1.0,
-                                            2.0,
+                                            (float)1.0,
+                                            (float)2.0,
                                             new Color(Color::red()),
                                             NULL,
-                                            1.0,
+                                            (float)1.0,
                                             false,
                                             NULL);
 
@@ -247,8 +247,8 @@ Mesh* AbstractMesh::createNormalsMesh() const{
                                            true,
                                            _center,
                                            fbb->create(),
-                                           2.0,
-                                           1.0,
+                                           (float)2.0,
+                                           (float)1.0,
                                            new Color(Color::blue()));
 
   delete fbb;
